feature:设置dto的字段类型；完成connect报文的实现；完成raft的改造

+++ server/src/main/java/com/ep/mqtt/server/db/dto/AsyncJobDto.java
@@ -2,6 +2,7 @@ package com.ep.mqtt.server.db.dto;
 
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
+
 import lombok.Data;
 
 /**
@@ -18,11 +19,11 @@ public class AsyncJobDto {
 
     private String businessType;
 
-    private Integer lastStartTime;
+    private Long lastStartTime;
 
-    private Integer lastEndTime;
+    private Long lastEndTime;
 
-    private Integer nextExecuteTime;
+    private Long nextExecuteTime;
 
     private Integer executeNum;
