+++ server/src/main/java/com/ep/mqtt/server/db/dto/SendMessageDto.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.ep.mqtt.server.metadata.Qos;
 import com.ep.mqtt.server.metadata.YesOrNo;
+
 import lombok.Data;
 
 /**
@@ -38,6 +39,6 @@ public class SendMessageDto {
 
     private YesOrNo isReceivePubcomp;
 
-    private Integer validTime;
+    private Long validTime;
 
 }
